Solution to Input Two Floats and Print Their Product

+++ Input Two Floats and Print Their Product@@6788d5012db25efec946f93e/code.c
@@ -3,7 +3,7 @@
 int main(){
     float x,y;
     scanf("%f",&x,&y);
-    int a=x*y;
+    float a=x*y;
     printf("Product: %f",a);
 
     return 0;
